Reorganized rkebootstrap_types.go
